feat!: version 0.5.0 release
Crate hierarchy is restructured for greater clarity in function of each
module and its submodules.

Sink implementation and playback has moved from rodio with the symphonia
backend to a custom sink implementation on top of symphonia. Reason: an
exercise in learning a little bit more about audio systems.

work in progress